HDDS-12055. Move ozone debug container to ozone debug datanode container
HDDS-12055. Move ozone debug container to ozone debug datanode container

Move ozone debug container to ozone debug datanode container

Fixed acceptance test

Modified name
